Separate internal tests from IRI tests; Test via public interface
